+++ Terraform/Azure/modules/dns/main.tf
@@ -1,7 +1,8 @@
 # Configure Azure DNS for Public name resolution.
 # progresolplus.pe
 resource "azurerm_dns_zone" "dns" {
-  name                = local.dns
+  for_each            = { for idx, dns in var.dns : idx => dns }
+  name                = "${tostring(each.value.name)}"
   resource_group_name = local.group
 
   soa_record {
@@ -10,8 +11,8 @@ resource "azurerm_dns_zone" "dns" {
       environment = local.environment
     }
 
-    host_name = "ns1-34.azure-dns.com"
-    email     = "azuredns-hostmaster.microsoft.com"
+    host_name = "${tostring(each.value.host_name)}"
+    email     = local.email
   }
 }
 
@@ -36,14 +37,14 @@ resource "azurerm_dns_a_record" "backoffice" {
   resource_group_name = local.group
   name                = "backoffice"
   records             = ["20.119.0.4"]
-  zone_name           = azurerm_dns_zone.dns.name
+  zone_name           = azurerm_dns_zone.dns[1].name
 }
 
 resource "azurerm_dns_a_record" "test" {
   ttl                 = local.ttl
   name                = local.test
   resource_group_name = local.group
-  zone_name           = azurerm_dns_zone.dns.name
+  zone_name           = azurerm_dns_zone.dns[1].name
   records             = ["20.172.245.13", "20.119.0.4"]
 }
 
@@ -56,7 +57,7 @@ resource "azurerm_dns_ns_record" "progresolplus" {
   ttl                 = local.ttl2
   resource_group_name = local.group
   name                = local.private2
-  zone_name           = azurerm_dns_zone.dns.name
+  zone_name           = azurerm_dns_zone.dns[1].name
   records             = ["ns1-34.azure-dns.com", "ns2-34.azure-dns.net", "ns3-34.azure-dns.org", "ns4-34.azure-dns.info"]
 }
 
@@ -69,7 +70,7 @@ resource "azurerm_dns_mx_record" "progresolplus" {
   ttl                 = local.ttl3
   resource_group_name = local.group
   name                = local.private2
-  zone_name           = azurerm_dns_zone.dns.name
+  zone_name           = azurerm_dns_zone.dns[1].name
 
   record {
     preference = 0
@@ -81,7 +82,7 @@ resource "azurerm_dns_mx_record" "progresolplus" {
 resource "azurerm_dns_cname_record" "backoffice-dev" {
   ttl                 = local.ttl
   resource_group_name = local.group
-  zone_name           = azurerm_dns_zone.dns.name
+  zone_name           = azurerm_dns_zone.dns[1].name
   name                = "backoffice-${local.environment}"
   record              = "backoffice-dev.progresolplus.pe.cdn.cloudflare.net"
 }
@@ -90,7 +91,7 @@ resource "azurerm_dns_cname_record" "backoffice-dev" {
 resource "azurerm_dns_cname_record" "front-dev" {
   ttl                 = local.ttl3
   resource_group_name = local.group
-  zone_name           = azurerm_dns_zone.dns.name
+  zone_name           = azurerm_dns_zone.dns[1].name
   name                = "front-${local.environment}"
   record              = "front-dev.progresolplus.pe.cdn.cloudflare.net"
 }
@@ -99,7 +100,7 @@ resource "azurerm_dns_cname_record" "storybook-dev" {
   ttl                 = local.ttl3
   resource_group_name = local.group
   record              = "storybook-dev"
-  zone_name           = azurerm_dns_zone.dns.name
+  zone_name           = azurerm_dns_zone.dns[1].name
 
   name                = "storybook-${local.environment}"
 }
@@ -128,7 +129,7 @@ resource "azurerm_dns_txt_record" "progresolplus-pe-asuid" {
 
   ttl                 = local.ttl
   resource_group_name = local.group
-  zone_name           = azurerm_dns_zone.dns.name
+  zone_name           = azurerm_dns_zone.dns[1].name
   name                = "progresolplus.pe.asuid.backoffice"
 
   record {
@@ -141,7 +142,7 @@ resource "azurerm_dns_cname_record" "autodiscover" {
   resource_group_name = local.group
   name                = "autodiscover"
   record              = "autodiscover.outlook.com"
-  zone_name           = azurerm_dns_zone.dns.name
+  zone_name           = azurerm_dns_zone.dns[1].name
 }
 
 resource "azurerm_dns_txt_record" "progresolplus-asuid" {
@@ -157,7 +158,7 @@ resource "azurerm_dns_txt_record" "progresolplus-asuid" {
   ttl                 = local.ttl
   resource_group_name = local.group
   name                = "asuid.backoffice"
-  zone_name           = azurerm_dns_zone.dns.name
+  zone_name           = azurerm_dns_zone.dns[1].name
 }
 
 resource "azurerm_dns_txt_record" "cloudflare-verify" {
@@ -173,7 +174,7 @@ resource "azurerm_dns_txt_record" "cloudflare-verify" {
   ttl                 = local.ttl
   resource_group_name = local.group
   name                = "cloudflare-verify"
-  zone_name           = azurerm_dns_zone.dns.name
+  zone_name           = azurerm_dns_zone.dns[1].name
 }
 
 resource "azurerm_dns_txt_record" "asuid-storybook-dev" {
@@ -184,7 +185,7 @@ resource "azurerm_dns_txt_record" "asuid-storybook-dev" {
 
   ttl                 = local.ttl3
   resource_group_name = local.group
-  zone_name           = azurerm_dns_zone.dns.name
+  zone_name           = azurerm_dns_zone.dns[1].name
   name                = "asuid.storybook-${local.environment}"
 
   record {
@@ -197,89 +198,13 @@ resource "azurerm_dns_a_record" "www" {
   ttl                 = local.ttl3
   resource_group_name = local.group
   records             = ["20.172.245.13"]
-  zone_name           = azurerm_dns_zone.dns.name
+  zone_name           = azurerm_dns_zone.dns[1].name
 }
-
-# progresolmas.com
-resource "azurerm_dns_zone" "dns2" {
-  name                = local.dns2
-  resource_group_name = local.group
-
-  soa_record {
-    tags = {
-      created_by  = local.created
-      environment = local.environment
-    }
-
-    host_name = "ns1-33.azure-dns.com"
-    email     = "azuredns-hostmaster.microsoft.com"
-  }
-}
-
-/* resource "azurerm_dns_ns_record" "progresolmas" {
-  name                = "@"
-  zone_name           = azurerm_dns_zone.dns2.name
-  resource_group_name = local.group
-  ttl                 = local.ttl2
-  records              = ["ns1-33.azure-dns.com", "ns2-33.azure-dns.net", "ns3-33.azure-dns.org", "ns4-33.azure-dns.info"]
-} */
-
-# progresolmas.pe
-resource "azurerm_dns_zone" "dns3" {
-  name                = local.dns3
-  resource_group_name = local.group
-
-  soa_record {
-    tags = {
-      created_by  = local.created
-      environment = local.environment
-    }
-
-    host_name = "ns1-35.azure-dns.com"
-    email     = "azuredns-hostmaster.microsoft.com"
-  }
-}
-
-/* resource "azurerm_dns_ns_record" "progresolmas-pe-ns" {
-  name                = "@"
-  zone_name           = azurerm_dns_zone.dns3.name
-  resource_group_name = local.group
-  ttl                 = local.ttl2
-  records              = ["ns1-35.azure-dns.com", "ns2-35.azure-dns.net", "ns3-35.azure-dns.org", "ns4-35.azure-dns.info"]
-} */
 
 resource "azurerm_dns_a_record" "progresolmas-test" {
   ttl                 = 10
   name                = local.test
   resource_group_name = local.group
   records             = ["10.0.1.2"]
-  zone_name           = azurerm_dns_zone.dns3.name
+  zone_name           = azurerm_dns_zone.dns[0].name
 }
-
-# progresolplus.com
-resource "azurerm_dns_zone" "dns4" {
-  name                = local.dns4
-  resource_group_name = local.group
-
-  soa_record {
-    tags = {
-      created_by  = local.created
-      environment = local.environment
-    }
-
-    host_name = "ns1-37.azure-dns.com"
-    email     = "azuredns-hostmaster.microsoft.com"
-  }
-}
-
-/* resource "azurerm_dns_ns_record" "progresolplus-com-ns" {
-  name                = "@"
-  zone_name           = azurerm_dns_zone.dns4.name
-  resource_group_name = local.group
-  ttl                 = local.ttl2
-  records              = ["ns1-37.azure-dns.com", "ns2-37.azure-dns.net", "ns3-37.azure-dns.org", "ns4-37.azure-dns.info"]
-
-  tags = {
-    environment = local.environment
-  }
-} */