+++ Terraform/Azure/modules/dns/variables.tf
@@ -1,8 +1,10 @@
+variable "dns" {
+  type = list(object({
+    name      = string
+    host_name = string
+  }))
+}
 
-variable "dns"         { type = string }
-variable "dns2"        { type = string }
-variable "dns3"        { type = string }
-variable "dns4"        { type = string }
 variable "group"       { type = string }
 variable "private"     { type = string }
 variable "environment" { type = string }
@@ -11,16 +13,13 @@ locals {
   ttl         = "300"
   test        = "test"
   ttl3        = "3600"
-  dns         = var.dns
-  dns2        = var.dns2
-  dns3        = var.dns3
-  dns4        = var.dns4
   ttl2        = "172800"
   group       = var.group
   private2    = "private"
   private     = var.private
   created     = "terraform"
   environment = var.environment
+  email       = "azuredns-hostmaster.microsoft.com"
   value       = "BE59ABCD5CA8C90D1A3AA0D6B37FCA7854E9A5AC000451FD8B8221B6B4C69B5D"
 }