+++ Terraform/Azure/modules/groups/output.tf
@@ -1,5 +1,5 @@
 # Outputs
-output "vms"        { value = azurerm_resource_group.vm.name }
-output "apps"       { value = azurerm_resource_group.apps.name }
-output "storage"    { value = azurerm_resource_group.storage.name }
-output "networking" { value = azurerm_resource_group.networking.name }
+output "vms"        { value = azurerm_resource_group.apps[0].name }
+output "apps"       { value = azurerm_resource_group.apps[1].name }
+output "storage"    { value = azurerm_resource_group.apps[2].name }
+output "networking" { value = azurerm_resource_group.apps[3].name }