+++ Terraform/Azure/modules/networking/output.tf
@@ -1,23 +1,12 @@
 # Outputs
-# ID PUBLIC SUBNET.
-output "webs" {
-  value = azurerm_subnet.web.id
+output "apps" { value = azurerm_virtual_network.apps }
+
+# ID PUBLIC SUBNETS.
+output "gateway" {
+  value = [for idx, gateway in azurerm_subnet.gateway : { index = idx, id = gateway.id }]
 }
 
-# ID PUBLIC SUBNET Virtual Machines.
-output "pvm" {
-  value = azurerm_subnet.pvm.id
-}
-
-output "functions" {
-  value = azurerm_subnet.function.id
-}
-
-# ID PUBLIC SUBNET APP Functions.
-output "pfunction" {
-  value = azurerm_subnet.pfunction.id
-}
-
-output "apps" {
-  value = azurerm_virtual_network.apps.id
+# ID SUBNETS.
+output "subnet" {
+  value = [for idx, subnet in azurerm_subnet.subnet : { index = idx, id = subnet.id }]
 }