+++ Terraform/Azure/modules/plan/main.tf
@@ -1,28 +1,16 @@
 #-------------------------------------
 # SPLAN Creation - Default is "true"
 #-------------------------------------
-resource "azurerm_service_plan" "webs" {
+resource "azurerm_service_plan" "apps" {
   tags = {
     created_by  = local.created
     environment = local.environment
   }
 
-  os_type             = local.os
-  sku_name            = local.sku
-  name                = local.web
-  resource_group_name = local.group
-  location            = local.location
-}
-
-resource "azurerm_service_plan" "functions" {
-  tags = {
-    created_by  = local.created
-    environment = local.environment
-  }
-
+  count               = length(var.app)
+  name                = "${local.name}-${var.app[count.index]}"
   os_type             = local.os
   sku_name            = local.sku
   resource_group_name = local.group
-  name                = local.function
   location            = local.location
 }