+++ Terraform/Azure/modules/plan/output.tf
@@ -1,8 +1,8 @@
 # Outputs
 output "web" {
-  value = azurerm_service_plan.webs.id
+  value = azurerm_service_plan.apps[0].id
 }
 
 output "function" {
-  value = azurerm_service_plan.functions.id
+  value = azurerm_service_plan.apps[1].id
 }