+++ Terraform/Azure/CHANGELOG.md
@@ -20,6 +20,33 @@ Initial release
 
 - Integrate: Add Environment Variable "ASPNETCORE_ENVIRONMENT" for Application Settings in every Linux Web App with DotNet & Node.
 - Integrate: Random String on Module Apps in Nanaykuna for nexts deploys.
-- Integrate: Add Environment Variable "API_KEY_SECRET" for Application Settings in every Linux Web App with DotNet & Node.
+![Alt text](image-5.png)
 
+- Integrate: Add Environment Variable "API_KEY_SECRET" for Application Settings in every Linux Web App with DotNet & Node.
+![Alt text](items.png)
+
+## Version [1.3.0]
+
+### Date: 19/07/2023
+
+- Update: Add a new Azure architecture for future and new deploys, about in use with functions length, variables as a type of list of strings by loops in count, with a index interpolation with local variables for Output in Resource.
 ![Alt text](image.png)
+
+![Alt text](image-2.png)
+
+![Alt text](image-3.png)
+
+![Alt text](image-5.png)
+
+- Update: Add a new Azure architecture for future and new deploys, about in use with function for, variables as a type of list of objects by loops for every property with type of string in foreach, with a index changed to string more interpolation with local variables for Output in Object and Resource.
+![Alt text](image-1.png)
+
+![Alt text](image-4.png)
+
+### Date: 21/07/2023
+
+- Update: Add a new Azure architecture for future and new deploys, about in use with function for, variables as a type of list of objects by loops for every property with type of string in foreach, with a index changed to string more interpolation with local variables for Output in Object and Resource for a new Resource or depend.
+
+![Alt text](items-3.png)
+
+![Alt text](items-2.png)